Maquina_virtual_2/backEnd/requirements.txt
fastapi
elasticsearch
uvicorn 
mysql.connector
python-dotenv


